requirements.txt
backoff>=2.2.1
pandas>=1.5.3
pytest>=7.2.1
python-dateutil>=2.9.0
python-dotenv>=0.19.0
requests>=2.32.0